Added Authorization

+++ backend/controllers/posts.js
@@ -41,7 +41,8 @@ exports.addPostController = (req, res, next) => {
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + '/images/' + req.file.filename
+        imagePath: url + '/images/' + req.file.filename,
+        creator: req.userData.userId
     });
     post.save().then(createdPost => {
         res.status(201).json({
@@ -86,12 +87,17 @@ exports.updatePostController = (req, res, next) => {
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     });
     console.log('Post from server: ' + post);
-    Post.updateOne({ _id: req.params.id }, post)
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
         .then(result => {
-            res.status(200).json({ message: 'Update Successful' });
+            if (result.n > 0) {
+                res.status(200).json({ message: 'Update Successful' });
+            } else {
+                res.status(401).json({ message: 'Not Authorized' });
+            }
         });
 };
 
@@ -103,9 +109,12 @@ exports.updatePostController = (req, res, next) => {
  */
 exports.deletePostController = (req, res, next) => {
     // console.log(req.params.id);
-    Post.deleteOne({ _id: req.params.id })
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
         .then(result => {
-            console.log(result);
-            res.status(200).json({ message: 'Post Deleted!' });
+            if (result.n > 0) {
+                res.status(200).json({ message: 'Post Deleted' });
+            } else {
+                res.status(401).json({ message: 'Not Authorized' });
+            }
         });
 };
